requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
whitenoise==6.5.0
gunicorn==21.2.0

